second change

+++ src/examples/MedicalTextIndexer.java
@@ -5,6 +5,7 @@ import gov.nih.nlm.nls.skr.GenericObject;
 import java.io.File;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.Arrays;
 
 public class MedicalTextIndexer {
     // 通过用户名、密码、邮箱、txt文件路径，返回结果
@@ -36,24 +37,39 @@ public class MedicalTextIndexer {
         return "";
     }
 
+    public String[] getFilenames(String inputPath) {
+        File file = new File(inputPath);
+        String[] fileNames = file.list();
+//        Arrays.sort(fileNames);
+//        for(int i = 0; i < fileNames.length; i++) {
+//            System.out.println(fileNames[i]);
+//        }
+        return fileNames;
+    }
+
     // 通过txt文件目录名批量处理
-    public void getResultsBatch(String inputPath, String outputPath) {
+    public void getResultsBatch(String inputPath, String outputPath, String[] fileNames, int start, int end) {
         String username = "";
         String password = "";
         String email = "";
 
-        File file = new File(inputPath);
-        String[] fileNames = file.list();
+//        File file = new File(inputPath);
+//        String[] fileNames = file.list();
 
-        for(int i = 0; i < fileNames.length; i++) {
+        for(int i = (start-1); i < (end-1); i++) {
             String filename = fileNames[i];
             Path inPath = Paths.get(inputPath, filename);
             Path outPath = Paths.get(outputPath, filename);
+            File outfile = new File(outPath.toString());
+            if (outfile.exists()) {
+                System.out.println(outPath.toString() + " exists! i = " + i);
+                continue;
+            }
             String results = getResults(username, password, email, inPath.toString());
             FileUtil fileUtil = new FileUtil();
             fileUtil.writeFile(results, outPath.toString());
 
-            System.out.println(outPath.toString() + " save success! i = " + i);
+            System.out.println(outPath.toString() + " save success! i = " + i + " ********");
         }
         System.out.println("All tags get success!");
     }
@@ -63,6 +79,7 @@ public class MedicalTextIndexer {
         String outputPath = "/Users/wanglei/Documents/IdeaProjects/MTI/data/tag10";
 
         MedicalTextIndexer mti = new MedicalTextIndexer();
-        mti.getResultsBatch(inputPath, outputPath);
+        String[] fileNames = mti.getFilenames(inputPath);
+        mti.getResultsBatch(inputPath, outputPath, fileNames, 0, 5);
     }
 }
